\name{GSQR-package}
\alias{GSQR-package}
\alias{GSQR}
\docType{package}
\title{
\packageTitle{GSQR}
}
\description{
\packageDescription{GSQR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GSQR}
\packageIndices{GSQR}

}
\author{
\packageAuthor{GSQR}

Maintainer: \packageMaintainer{GSQR}
}
\references{
Mohamed Ouhourane, Yi Yang, Mohamed Ouhourane "Group Penalized Smooth Quantile Regression"
}


